Segmented toneline drawn but not accurate characteristics
Much cleaner drawing function which is generalised for tonlines
consisting of multiple line segments. The second tone now behaves better
but the second and the fourth tone has become worse. Need to get cleaner
lines for third and fourth tone before #2 is fixed. Might have to do with the
dynamic plot range too.