Separate spectrogramModel
The process of refactoring the code into a structure with one model for
spectrogramoutput and one for toneanalysis has started. #14 still needs a new
version of the toneModel which takes a input spectrogram from the spectroGramModel
and outputs a toneline to the view. Should the same instances of the inputStates
be accessed by all the models and view? Should there be a mediatior that communicate
state information between the different views and models?