Soundfile playback after ended fixed!
It did not work properly to set this._audio.currentTime to 0 when
I wanted to reset the audio file. Finally found this hack:
http://stackoverflow.com/questions/16966724/html5-audio-currenttime-doesnt-work
I think it might have had to do with my file being a .wav file in an
immature stage of the Web Audio API development. Halleluja, this fixes #1 !