Highpass convolution on spectrogram
To increase the contrast in the spectrogram it is now gone through
a convolution with a highpass filter. This is practically a lateral
inhibitory network and works well to enchance the peaks. The result is that
the HPS algorithm works a bit better. This fixes #18. Also fixes #13 since not much more
can be done to improve the HPS result except if a higher resolution of
the spectrogram was possible which it is not atm.