Delayed starpower drawing with sound effects for each star
To play the same audio element multiple times currently results in a new HTTP
each time. According to [this](http://stackoverflow.com/questions/10385779/in-html5-can-i-play-the-same-sound-more-than-once-at-the-same-time)
post this seems to be unavoidable atm. Still a memory heap shows (in incognito mode)
that no extra data is stored each time so fortunately the GC seems to remove
things as it should.